+++ R/corpus_reshape.R
@@ -5,10 +5,11 @@
 #' Because the corpus object records its current "units" status, it is possible
 #' to move from recast units back to original units, for example from documents,
 #' to sentences, and then back to documents (possibly after modifying the sentences).
-#' @param x corpus whose document units will be reshaped
-#' @param to new document units in which the corpus will be recast
+#' @param x corpus whose document units will be reshaped.
+#' @param to new document units in which the corpus will be recast.
+#' @param verbose if `TRUE`, print timing messages to the console,
 #' @param ... additional arguments passed to [tokens()], since the
-#'   syntactic segmenter uses this function)
+#'   syntactic segmenter uses this function).
 #' @inheritParams corpus_segment
 #' @return A corpus object with the documents defined as the new units,
 #'   including document-level meta-data identifying the original documents.
@@ -30,24 +31,27 @@
 #' @export
 #' @keywords corpus
 corpus_reshape <- function(x, to = c("sentences", "paragraphs", "documents"),
-                           use_docvars = TRUE, ...) {
+                           use_docvars = TRUE, verbose = FALSE, ...) {
     UseMethod("corpus_reshape")
 }
 
 #' @export
 corpus_reshape.default <- function(x, to = c("sentences", "paragraphs", "documents"),
-                                   use_docvars = TRUE, ...) {
+                                   use_docvars = TRUE, verbose = FALSE, ...) {
     check_class(class(x), "corpus_reshape")
 }
 
 #' @export
 corpus_reshape.corpus <- function(x, to = c("sentences", "paragraphs", "documents"),
-                                  use_docvars = TRUE, ...) {
+                                  use_docvars = TRUE, verbose = FALSE, ...) {
 
     x <- as.corpus(x)
     to <- match.arg(to)
     use_docvars <- check_logical(use_docvars)
     
+    if (verbose) 
+        catm("Reshaping corpus...\n")
+        
     attrs <- attributes(x)
     if (field_object(attrs, "unit") == to)
         return(x)
@@ -65,7 +69,7 @@ corpus_reshape.corpus <- function(x, to = c("sentences", "paragraphs", "document
         unit <- "documents"
     } else {
         temp <- segment_texts(x,  pattern = NULL, extract_pattern = FALSE,
-                              omit_empty = FALSE, what = to, ...)
+                              omit_empty = FALSE, what = to, verbose = verbose, ...)
         attrs[["docvars"]] <- reshape_docvars(attrs[["docvars"]], temp[["docnum"]], unique = FALSE)
         unit <- to
     }